Revert "[feat(metrics)] Metric API 서버 구현"
