+++ backend/docker-elasticsearch/compose.yml
@@ -1,0 +1,18 @@
+services:
+  elastic:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - 9200:9200 # 9200번 포트에서 Elasticsearch 실행
+    environment:
+      # 아래 설정은 개발/테스트 환경에서 간단하게 테스트하기 위한 옵션 (운영 환경에서는 설정하면 안 됨.)
+      - discovery.type=single-node # 단일 노드
+      - xpack.security.enabled=false # 보안 설정
+      - xpack.security.http.ssl.enabled=false # 보안 설정
+  kibana:
+    image: docker.elastic.co/kibana/kibana:8.17.4
+    ports:
+      - 5601:5601 # 5601번 포트에서 kibana 실행
+    environment:
+      - ELASTICSEARCH_HOSTS=http://elastic:9200 # kibana가 통신할 Elasticsearch
