+++ backend/src/app.module.ts
@@ -1,10 +1,9 @@
 import { Module } from "@nestjs/common";
 import { AppController } from "./app.controller";
 import { LogModule } from "./logs/logs.module";
-import { MetricsModule } from "./metrics/metrics.module";
 
 @Module({
-  imports: [LogModule, MetricsModule],
+  imports: [LogModule],
   controllers: [AppController],
   providers: [],
 })
