+++ backend/src/log-consumer/log-consumer.main.ts
@@ -1,9 +1,6 @@
-import { config } from "dotenv";
-config(); // Load .env file before anything else
-
 import { NestFactory } from "@nestjs/core";
 import { MicroserviceOptions, Transport } from "@nestjs/microservices";
-import { LogConsumerModule } from "./event-consumer.module";
+import { LogConsumerModule } from "./log-consumer.module";
 
 function parseBrokers(): string[] {
   const brokersEnv = process.env.KAFKA_BROKERS;
