+++ backend/src/log-consumer/log-consumer.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+import { LogModule } from "../logs/logs.module";
+import { LogConsumer } from "./log.consumer";
+
+@Module({
+  imports: [LogModule],
+  controllers: [LogConsumer],
+})
+export class LogConsumerModule {}
