--- infra/docker-compose.yml
@@ -1,98 +1,0 @@
-services:
-  # Kafka - 메시지 큐
-  kafka:
-    image: apache/kafka:3.7.0
-    container_name: panopticon-kafka
-    environment:
-      KAFKA_NODE_ID: 1
-      KAFKA_PROCESS_ROLES: broker,controller
-      KAFKA_CONTROLLER_QUORUM_VOTERS: 1@kafka:9093
-      KAFKA_CONTROLLER_LISTENER_NAMES: CONTROLLER
-      KAFKA_LISTENERS: PLAINTEXT://:9092,CONTROLLER://:9093
-      KAFKA_ADVERTISED_LISTENERS: PLAINTEXT://localhost:9092
-      KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: CONTROLLER:PLAINTEXT,PLAINTEXT:PLAINTEXT
-      KAFKA_LOG_DIRS: /var/lib/kafka/data
-      KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
-      KAFKA_TRANSACTION_STATE_LOG_REPLICATION_FACTOR: 1
-      KAFKA_TRANSACTION_STATE_LOG_MIN_ISR: 1
-    ports:
-      - "9092:9092"
-    volumes:
-      - kafka-data:/var/lib/kafka/data
-    networks:
-      - panopticon-network
-    restart: unless-stopped
-
-  # TimescaleDB - 메트릭 저장소 (PostgreSQL 기반)
-  timescaledb:
-    image: timescale/timescaledb:latest-pg16
-    container_name: panopticon-timescaledb
-    environment:
-      POSTGRES_DB: panopticon
-      POSTGRES_USER: admin
-      POSTGRES_PASSWORD: admin123
-    ports:
-      - "5433:5432"
-    volumes:
-      - timescale-data:/var/lib/postgresql/data
-      - ../timescaledb/init_metrics_tables.sql:/docker-entrypoint-initdb.d/init_metrics_tables.sql
-      - ../timescaledb/init_timescale_test.sql:/docker-entrypoint-initdb.d/init_timescale_test.sql
-    networks:
-      - panopticon-network
-    restart: unless-stopped
-
-  # Elasticsearch - 로그 저장소
-  elasticsearch:
-    build:
-      context: ./elasticsearch
-      dockerfile: Dockerfile
-    container_name: panopticon-elasticsearch
-    ports:
-      - "9200:9200"
-    environment:
-      - discovery.type=single-node
-      - xpack.security.enabled=false
-      - xpack.security.http.ssl.enabled=false
-      - "ES_JAVA_OPTS=-Xms512m -Xmx512m"
-    volumes:
-      - elasticsearch-data:/usr/share/elasticsearch/data
-    networks:
-      - panopticon-network
-    restart: unless-stopped
-
-  # Kibana - Elasticsearch UI
-  kibana:
-    image: docker.elastic.co/kibana/kibana:8.17.4
-    container_name: panopticon-kibana
-    ports:
-      - "5601:5601"
-    environment:
-      - ELASTICSEARCH_HOSTS=http://elasticsearch:9200
-    depends_on:
-      - elasticsearch
-    networks:
-      - panopticon-network
-    restart: unless-stopped
-
-  # Redis - 실시간 집계 캐시
-  redis:
-    image: redis:7-alpine
-    container_name: panopticon-redis
-    ports:
-      - "6379:6379"
-    volumes:
-      - redis-data:/data
-    networks:
-      - panopticon-network
-    restart: unless-stopped
-    command: redis-server --appendonly yes
-
-volumes:
-  kafka-data:
-  timescale-data:
-  elasticsearch-data:
-  redis-data:
-
-networks:
-  panopticon-network:
-    driver: bridge
